+++ history.md
@@ -1,6 +1,7 @@
 ### History
 
 v2.8.0
+- Allow adaptive precision in data range spinboxes.
 - Enhanced map settings panel.
 - Added distinct color assignment to record list context menu.
 - Added time series rename actions using F2 and context menu.